Server/requirements.txt
absl-py==0.13.0
alembic==1.6.5
astor==0.8.1
astroid==2.3.3
cached-property==1.5.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.3
dnspython==2.1.0
docopt==0.6.2
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gast==0.5.0
grpcio==1.38.1
gunicorn==20.0.4
h5py==3.3.0
hbmqtt==0.9.6
idna==2.10
importlib-metadata==4.6.1
isort==4.3.21
itsdangerous==2.0.1
Jinja2==2.11.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mock==4.0.3
mysqlclient==2.0.3
numpy==1.21.0
opencv-python==4.5.2.54
paho-mqtt==1.5.1
passlib==1.7.4
protobuf==3.17.3
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.4.4
pyserial==3.5
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.18.0
python-editor==1.0.4
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.3.23
stcgal==1.6
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
tqdm==4.55.0
transitions==0.8.8
typed-ast==1.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
websockets==9.1
Werkzeug==2.0.1
wincertstore==0.2
WTForms==2.3.3
zipp==3.5.0
